[카드 추천] 🐿 유저별로 카드 추천 로직

+++ src/main/java/com/example/PayAll_BE/entity/Benefit.java
@@ -36,6 +36,6 @@ public class Benefit {
 	private Product product;
 
 	@Column(nullable = false)
-	private BigDecimal benefitValue; // 혜택 비율 (예: 10.00 -> 10%)
+	private Long benefitValue; // 혜택 비율 (예: 10.00 -> 10%)
 
 }
